refactor: constant 분리
